% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_temp_precip_plot.R
\name{plot_one_site}
\alias{plot_one_site}
\title{Plot one site}
\usage{
plot_one_site(site_clicked, site_df)
}
\arguments{
\item{site_clicked}{the site_id returned by clicking the map}

\item{site_df}{the data-frame of downloaded sites}
}
\value{
a two-panel plat
}
\description{
Plot one site
}
