% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_temp_precip_plot.R
\name{plot_site_env}
\alias{plot_site_env}
\title{Plot environmental  variables at at site}
\usage{
plot_site_env(type, site)
}
\arguments{
\item{type}{type of environment. one of "precip" or "temp"}

\item{site}{three digit code -- is probably actually the CELL id}
}
\value{

}
\description{
Plot environmental  variables at at site
}
