% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_region_join.R
\name{site_region_join}
\alias{site_region_join}
\title{join ouranos regions and the downloaded sites}
\usage{
site_region_join(site_sf_df)
}
\arguments{
\item{site_sf_df}{dataframe from [rcoleo::download_sites_sf()]}
}
\value{

}
\description{
join ouranos regions and the downloaded sites
}
