% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_temp_precip_plot.R
\name{plot_one_site}
\alias{plot_one_site}
\title{Plot one site}
\usage{
plot_one_site(site_clicked, site_df, lookup_vec)
}
\arguments{
\item{site_clicked}{the site display_name returned by clicking the map}

\item{site_df}{the data-frame of downloaded sites}

\item{lookup_vec}{a vector of lookup values. should take cell_ids and map them to something.}
}
\value{
a two-panel plat
}
\description{
This is a BADLY NAMED FUNCTION -- it plots a cell, not a site!
}
